+++ timelock-server-benchmark-client/src/main/java/com/palantir/atlasdb/timelock/benchmarks/benchmarks/TransactionWriteRowsBenchmark.java
@@ -18,11 +18,11 @@ package com.palantir.atlasdb.timelock.benchmarks.benchmarks;
 import java.util.List;
 import java.util.Map;
 
-import com.palantir.atlasdb.timelock.benchmarks.RandomBytes;
 import com.palantir.atlasdb.timelock.benchmarks.schema.generated.BlobsTable;
 import com.palantir.atlasdb.timelock.benchmarks.schema.generated.BlobsTable.BlobsRow;
 import com.palantir.atlasdb.transaction.api.Transaction;
 import com.palantir.atlasdb.transaction.api.TransactionManager;
+import com.palantir.common.random.RandomBytes;
 
 public final class TransactionWriteRowsBenchmark extends AbstractWriteTransactionBenchmark {